Factor out common code
