+++ src/main.rs
@@ -24,6 +24,8 @@ mod resolve;
 mod rversion;
 mod utils;
 
+use crate::common::*;
+
 #[cfg(any(target_os = "macos", target_os = "linux"))]
 mod escalate;
